update to correct type

+++ libs/openFrameworks/video/ofAVFoundationGrabber.h
@@ -33,7 +33,7 @@ class ofAVFoundationGrabber;
 	int height;
 
 	BOOL bInitCalled;
-	int deviceID;
+	NSUInteger deviceID;
 
 	AVCaptureDeviceInput		*captureInput;
 	AVCaptureVideoDataOutput    *captureOutput;
